pyproject.toml
[project]
name = "wxpdfreader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.2",
    "nuitka>=2.8.9",
    "numpy>=2.3.5",
    "pymupdf==1.23.8",
    "pyopengl>=3.1.10",
    "wxpython>=4.2.4",
]
